refactor: update InvalidView import path to new location
