+++ src/app/webpay-mall-diferido/refund/page.tsx
@@ -6,7 +6,8 @@ import { getRefundTRXSteps } from "@/app/webpay-mall-diferido/content/steps/refu
 import { NextPageProps } from "@/types/general";
 import { refundTransaction } from "@/app/lib/webpay-mall-diferido/data";
 import { StatusButton } from "@/app/webpay-mall-diferido/components/StatusButton";
-import { InvalidView } from "@/app/invalid/invalidRefund";
+import { InvalidView } from "@/components/customError/CustomError";
+import { Metadata } from "next";
 
 const actualBread: Route[] = [
   {
